+++ Scorpio/src/Instruction/ScriptInstruction.cs
@@ -6,11 +6,11 @@ namespace Scorpio.Instruction {
     [StructLayout(LayoutKind.Sequential, Pack = 1)]
     public struct ScriptInstruction {
         public int opvalue;         //指令值
-        public int line;            //代码在多少行
+        public ushort line;         //代码在多少行
         public OpcodeType optype;   //指令类型
         public Opcode opcode;       //指令类型
-        public ScriptInstruction(int opcode, int opvalue, int line) : this((Opcode)opcode, opvalue, line) { }
-        public ScriptInstruction(Opcode opcode, int opvalue, int line) {
+        public ScriptInstruction(byte opcode, int opvalue, ushort line) : this((Opcode)opcode, opvalue, line) { }
+        public ScriptInstruction(Opcode opcode, int opvalue, ushort line) {
             this.optype = OpcodeType.Nop;
             this.opcode = opcode;
             this.opvalue = opvalue;
